chore: update COMPATIBLE_KESTRA_VERSION.properties to list all versions

+++ run-e2e-tests.sh
@@ -8,30 +8,37 @@ log_and_run() {
 }
 
 if [ $# -ge 1 ]; then
-  KESTRA_VERSION="$1"
+  versions="$1"
 else
-  KESTRA_VERSION=$(cat ./COMPATIBLE_KESTRA_VERSION.properties)
+  versions=$(cat ./COMPATIBLE_KESTRA_VERSION.properties)
 fi
 
 echo "/n------------------------------------------------"
 echo "Build Kestra CLI and test it with a docker Kestra instance"
-echo "docker KESTRA_VERSION used: $KESTRA_VERSION\n"
 
-echo "start Kestra container"
-log_and_run docker compose -f docker-compose-ci.yml down
+for KESTRA_VERSION in $versions; do
+  if [ -z "$KESTRA_VERSION" ]; then
+    continue
+  fi
 
-export KESTRA_VERSION=$KESTRA_VERSION
-log_and_run docker compose -f docker-compose-ci.yml up -d --wait || {
-   echo "db Docker Compose failed. Dumping logs:";
-   log_and_run docker compose -f docker-compose-ci.yml logs;
-   exit 1;
-}
+  echo "docker KESTRA_VERSION used: $KESTRA_VERSION\n"
 
-echo "build"
-log_and_run sh -c 'go build -C ./e2e_tests ./...'
+  echo "start Kestra container"
+  log_and_run docker compose -f docker-compose-ci.yml down
 
-echo "start tests"
-log_and_run sh -c 'go test -C ./e2e_tests ./...'
+  export KESTRA_VERSION=$KESTRA_VERSION
+  log_and_run docker compose -f docker-compose-ci.yml up -d --wait || {
+     echo "db Docker Compose failed. Dumping logs:";
+     log_and_run docker compose -f docker-compose-ci.yml logs;
+     exit 1;
+  }
 
-echo "stop Kestra container"
-log_and_run docker compose -f docker-compose-ci.yml down
+  echo "build"
+  log_and_run sh -c 'go build -C ./e2e_tests ./...'
+
+  echo "start tests"
+  log_and_run sh -c 'go test -C ./e2e_tests ./...'
+
+  echo "stop Kestra container"
+  log_and_run docker compose -f docker-compose-ci.yml down
+done
